Changes to frontend and draft of architecture chapter

+++ nipen/src/main/java/no/helsenorge/nipen/model/Weight.java
@@ -11,7 +11,7 @@ public class Weight extends DataType {
         this.unit = unit;
     }
 
-    public long getValue() {
+    public int getValue() {
         return value;
     }
